# backend-node/swagger-output.json
{
  "swagger": "2.0",
  "info": {
    "title": "Pharmalogy API",
    "description": "Complete API documentation for the Pharmalogy Pharmacy Management System backend.",
    "version": "2.0.0"
  },
  "host": "localhost:8001",
  "basePath": "/",
...
  "schemes": [
    "http",
    "https"
  ],
  "securityDefinitions": {
    "BearerAuth": {
      "type": "apiKey",
      "in": "header",
      "name": "Authorization",
      "description": "Please enter your JWT token provided by login endpoint. Example: \"Bearer eyJhbG...\""
    }
  },
  "paths": {
...
    "/api/bills/": {
      "get": {
        "description": "",
        "parameters": [
          {
            "name": "search",
            "in": "query",
            "type": "string"
          },
          {
            "name": "is_paid",
            "in": "query",
            "type": "string"
          },
          {
            "name": "start_date",
            "in": "query",
            "type": "string"
          },
          {
            "name": "end_date",
            "in": "query",
            "type": "string"
          },
          {
            "name": "page",
            "in": "query",
            "type": "string"
          },
          {
            "name": "limit",
            "in": "query",
            "type": "string"
          },
          {
            "name": "sort_by",
            "in": "query",
            "type": "string"
          },
          {
            "name": "sort_order",
            "in": "query",
            "type": "string"
          },
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          }
        }
      },
      "post": {
        "description": "",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "customer_id": {
                  "example": "any"
                },
                "customer_name": {
                  "example": "any"
                },
                "customer_mobile": {
                  "example": "any"
                },
                "customer_email": {
                  "example": "any"
                },
                "items": {
                  "example": "any"
                },
                "discount_percent": {
                  "example": "any"
                },
                "notes": {
                  "example": "any"
                },
                "is_paid": {
                  "example": "any"
                },
                "due_date": {
                  "example": "any"
                },
                "billing_date": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Created"
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          }
        }
      }
    },
    "/api/bills/insights": {
      "get": {
        "description": "",
        "parameters": [
          {
            "name": "start_date",
            "in": "query",
            "type": "string"
          },
          {
            "name": "end_date",
            "in": "query",
            "type": "string"
          },
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          }
        }
      }
    },
    "/api/bills/{bill_id}": {
      "get": {
        "description": "",
        "parameters": [
          {
            "name": "bill_id",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        }
      },
      "put": {
        "description": "",
        "parameters": [
          {
            "name": "bill_id",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "customer_id": {
                  "example": "any"
                },
                "customer_name": {
                  "example": "any"
                },
                "customer_mobile": {
                  "example": "any"
                },
                "customer_email": {
                  "example": "any"
                },
                "items": {
                  "example": "any"
                },
                "discount_percent": {
                  "example": "any"
                },
                "notes": {
                  "example": "any"
                },
                "is_paid": {
                  "example": "any"
                },
                "due_date": {
                  "example": "any"
                },
                "billing_date": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        }
      },
      "delete": {
        "description": "",
        "parameters": [
          {
            "name": "bill_id",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "restore_inventory",
            "in": "query",
            "type": "string"
          },
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
    "/api/bills/{bill_id}/mark-paid": {
      "post": {
        "description": "",
        "parameters": [
          {
            "name": "bill_id",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
    "/api/bills/{bill_id}/pdf": {
      "post": {
        "description": "",
        "parameters": [
          {
            "name": "bill_id",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
    "/api/bills/{bill_id}/email": {
      "post": {
        "description": "",
        "parameters": [
          {
            "name": "bill_id",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
...
  },
  "security": [
    {
      "BearerAuth": []
    }
  ]
}